Remove token from every parameters table
For the sync API the description of "Parameters" are related to the
command arguments, not API parameters.

This commit changes "Parameters" to "Command arguments" when it makes
sense and also removes token from every request. We added a not at the
the authorization section specifying that token is required
for **all** our resources.